refact: move definition to cpp
Move definition of qw_input_device::create to cpp.cpp
