+++ src/types/qwinputdevice.h
@@ -32,7 +32,7 @@ class QW_CLASS_OBJECT(input_device)
     Q_OBJECT
 
 public:
-    inline static qw_input_device *create(HandleType *handle);
+    static qw_input_device *create(HandleType *handle);
 
     QW_FUNC_MEMBER(input_device, is_wl, bool)
 #ifdef WLR_HAVE_X11_BACKEND
@@ -140,33 +140,4 @@ class QW_EXPORT qw_touch : public qw_input_device
     QW_SIGNAL(cancel, wlr_touch_cancel_event*)
     QW_SIGNAL(frame)
 };
-
-qw_input_device *qw_input_device::create(HandleType *handle) {
-    if (auto o = get(handle))
-        return o;
-
-    switch (handle->type) {
-    case WLR_INPUT_DEVICE_KEYBOARD:
-        return new qw_keyboard(handle, false);
-    case WLR_INPUT_DEVICE_POINTER:
-        return new qw_pointer(handle, false);
-#if WLR_VERSION_MINOR > 17
-    case WLR_INPUT_DEVICE_TABLET:
-#else
-    case WLR_INPUT_DEVICE_TABLET_TOOL:
-#endif
-        return new qw_tablet(handle, false);
-    case WLR_INPUT_DEVICE_TABLET_PAD:
-        return new qw_tablet_pad(handle, false);
-    case WLR_INPUT_DEVICE_SWITCH:
-        return new qw_switch(handle, false);
-    case WLR_INPUT_DEVICE_TOUCH:
-        return new qw_touch(handle, false);
-    default:
-        // Here is not reachable
-        qCritical("Unknow input device type!");
-        return nullptr;
-    }
-}
-
 QW_END_NAMESPACE
